+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 * System
   * Upgrade from 32 bit raspbian Bookworm Desktop to 64 bit raspbian Trixie Lite
   * Upgrade from Logitech Media Server 8.5.2 to Lyrion Music Server 9.0.3
+  * Upgraded volume calculations to preserve relative positions when hitting the min or max setting via source volume bar
+  * Added volume matching between AmpliPi and Spotify and vice-versa
+
+* Web App
+  * Changed caching rules to ensure that users don't get stuck with old versions of the webapp post update
 
 
 ## 0.4.10